requirements.txt
faster-whisper==0.9.0
websockets
onnxruntime==1.16.0
ffmpeg-python
scipy
websocket-client
tiktoken==0.3.3
kaldialign
braceexpand
openai-whisper
whisperspeech
soundfile

# CPU-mode dependencies
openai>=1.0.0
transformers>=4.30.0
sentencepiece
pyttsx3
